+++ backend/src/main/java/com/kielakjr/sports_events/service/SportService.java
@@ -1,5 +1,6 @@
 package com.kielakjr.sports_events.service;
 
+import com.kielakjr.sports_events.repo.TeamRepo;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,9 +13,10 @@ import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class SportService {
-
   @Autowired
   private SportRepo sportRepo;
+  @Autowired
+  private TeamRepo teamRepo;
 
   public List<Sport> getAllSports() {
     return sportRepo.findAll();
@@ -39,6 +41,9 @@ public class SportService {
   public void deleteSport(Long id) {
     if (!sportRepo.existsById(id)) {
       throw new EntityNotFoundException("Sport not found with id: " + id);
+    }
+    if (!teamRepo.findBySportId(id).isEmpty()) {
+      throw new IllegalStateException("Cannot delete sport with associated teams");
     }
     sportRepo.deleteById(id);
   }
